// slms-frontend/src/app/shared/models/user.model.ts
export type UserRole = 'Admin' | 'Captain';

export interface User {
  id?: number;
  name: string;
  email: string;
  role: UserRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  role: UserRole;
}

export interface AuthResponse {
  token: string;
  user?: User;
}